requirements.txt
PyQt6==6.9.0
pyperclip==1.8.2
python-xlib==0.33
tesserocr==2.6.0
Pillow==10.2.0 